The Last Dance

+++ GuideMe/graph.h
@@ -28,7 +28,7 @@ public:
 
     unordered_map<string, vector<Edge>> adj;// hashtable as every cell contains vector of edges , and adjacency lists
     map<int,vector<Edge>> road;
-    unordered_map<string,float>cost; //Cost for Dijkstra
+    unordered_map<string,int>cost; //Cost for Dijkstra
     map<pair<string,string>,set<pair<string,int>>>to_save;
 
     void getAllPaths(string s, string d,int b); // this fun choose which algorithms to use
